chore: remove old layout
