+++ src/pages/data-privacy.tsx
@@ -3,7 +3,7 @@ import SEO from '../components/seo';
 import { graphql } from 'gatsby';
 import { MarkdownAst } from '../components/markdown/markdown-ast';
 import { useTranslation } from 'gatsby-plugin-react-i18next';
-import { LayoutV2 } from '../components/layout/layout-v2';
+import { LayoutV2 } from '../components/layout/layout';
 import { SectionHeader } from '../new-components/section-header/section-header';
 import { LocalesQuery } from '../types';
 import { ContentBlockContainer } from '../components/layout/content-block-container';
