+++ src/pages/imprint.tsx
@@ -5,7 +5,7 @@ import styled from 'styled-components';
 import { Text } from '../components/typography/typography';
 import { MarkdownAst } from '../components/markdown/markdown-ast';
 import { useTranslation } from 'gatsby-plugin-react-i18next';
-import { LayoutV2 } from '../components/layout/layout-v2';
+import { LayoutV2 } from '../components/layout/layout';
 import { SectionHeader } from '../new-components/section-header/section-header';
 import { LocalesQuery } from '../types';
 import { ContentBlockContainer } from '../components/layout/content-block-container';
